tty: serial: owl: Reduce log spam when UART has no DMA support
Currently the kernel spams the logs during system boot and
shutdown. This happens because the init system continually
opens and closes the serial port when it logs a message which
causing the startup callback to continually attempt to open a
DMA channel that doesn't exist.

This is fixed by providing an early exit path that doesn't issue
any kernel messages.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>