NOUPSTREAM: deb-pkg: add bindeb-pkg as an alias
We are currently getting build failures because other code is starting
to use bindeb-pkg which was introduced in v4.1 by 3716001bcb7f
("deb-pkg: add source package").

Whilst this could be backported it requires some effort to disentangle
it from other related changes. For that reason I've taken a simpler
approach of aliasing bindeb-pkg and deb-pkg. This is OK because
bindeb-pkg in v4.1+ does the same thing that deb-pkg does in v3.10
anyway.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>